Move PublicDatasetTestUtils to o.a.h.fs.s3a.test
